==> 09_voronoi.py <==
#!/usr/bin/env python3
# -*- coding: utf-8 -*-

import numpy as np
import matplotlib.pyplot as plt
from scipy.spatial import Voronoi
import pandas as pd
from scipy.cluster.hierarchy import linkage, cut_tree
from scipy.spatial import voronoi_plot_2d
import matplotlib.cm as cm
import matplotlib as mpl
from sklearn.manifold import TSNE
from sklearn.preprocessing import StandardScaler
import scipy.stats as ss

def voronoi_finite_polygons_2d(vor, radius=None):

    if vor.points.shape[1] != 2:
        raise ValueError("Requires 2D input")

    new_regions = []
    new_vertices = vor.vertices.tolist()

    center = vor.points.mean(axis=0)
    if radius is None:
        radius = vor.points.ptp().max()*2

    # Construct a map containing all ridges for a given point
    all_ridges = {}
    for (p1, p2), (v1, v2) in zip(vor.ridge_points, vor.ridge_vertices):
        all_ridges.setdefault(p1, []).append((p2, v1, v2))
        all_ridges.setdefault(p2, []).append((p1, v1, v2))

    # Reconstruct infinite regions
    for p1, region in enumerate(vor.point_region):
        vertices = vor.regions[region]

        if all(v >= 0 for v in vertices):
            # finite region
            new_regions.append(vertices)
            continue

        # reconstruct a non-finite region
        ridges = all_ridges[p1]
        new_region = [v for v in vertices if v >= 0]

        for p2, v1, v2 in ridges:
            if v2 < 0:
                v1, v2 = v2, v1
            if v1 >= 0:
                # finite ridge: already in the region
                continue

            # Compute the missing endpoint of an infinite ridge

            t = vor.points[p2] - vor.points[p1] # tangent
            t /= np.linalg.norm(t)
            n = np.array([-t[1], t[0]])  # normal

            midpoint = vor.points[[p1, p2]].mean(axis=0)
            direction = np.sign(np.dot(midpoint - center, n)) * n
            far_point = vor.vertices[v2] + direction * radius

            new_region.append(len(new_vertices))
            new_vertices.append(far_point.tolist())

        # sort region counterclockwise
        vs = np.asarray([new_vertices[v] for v in new_region])
        c = vs.mean(axis=0)
        angles = np.arctan2(vs[:,1] - c[1], vs[:,0] - c[0])
        new_region = np.array(new_region)[np.argsort(angles)]

        # finish
        new_regions.append(new_region.tolist())

    return new_regions, np.asarray(new_vertices)



df = pd.read_csv("cbe_matrix14s.csv", index_col=0)
df = df.div(df.sum(axis=1), axis=0)

Z = linkage(df, method='single', metric='cosine')
cutree = cut_tree(Z, n_clusters=91)
df['CLUST'] = cutree

# create centroid matrix
dfg = df.groupby(['CLUST']).mean()
dfind = pd.DataFrame(columns=['sid','CLUST'])
dfind['sid'] = df.index
dfind['CLUST'] = df['CLUST'].to_list()
dfs = dfind.groupby(['CLUST'])['sid'].agg(lambda column: ",".join(column))
dfg['sid'] = dfs
dfg['nclust'] = dfg['sid'].apply(lambda x: x.count(',')+1)

# TSNE
X = dfg.iloc[: , :-2].to_numpy()
X = StandardScaler().fit_transform(X)
tsne = TSNE(n_components=2, init='pca', metric='cosine', square_distances=True, random_state=2) 
X2d = tsne.fit_transform(X)

# create voronoi points dataframe
dfpoints = pd.DataFrame(columns=['clid','x','y','sid','nclust'])
dfpoints['clid'] = dfg.index
dfpoints['sid'] = dfg['sid']
dfpoints['nclust'] = dfg['nclust']


nrank = ss.rankdata(dfg['nclust'])
minima = min(nrank)
maxima = max(nrank)
norm = mpl.colors.Normalize(vmin=minima, vmax=maxima, clip=True)
mapper = cm.ScalarMappable(norm=norm, cmap=cm.YlOrRd)

svgpol = ""
vidbox = ""
# create voronoi diagram
vor = Voronoi(X2d)
fig = voronoi_plot_2d(vor, show_vertices=False, line_colors='black', line_width=1.3, line_alpha=0.6, point_size=4)
regions, vertices = voronoi_finite_polygons_2d(vor)

xmin = ymin = float('inf')
xmax = ymax = float('-inf')
for pi in range(len(X2d)):
    if X2d[pi][0] < xmin:
        xmin = X2d[pi][0]
    if X2d[pi][0] > xmax:
        xmax = X2d[pi][0]
    if X2d[pi][1] < ymin:
        ymin = X2d[pi][1]
    if X2d[pi][1] > ymax:
        ymax = X2d[pi][1]

ri = 0
for region in regions:
    polygon = vertices[region]
    plt.fill(*zip(*polygon), alpha=0.4)
    pts = ""
    for vert in region:
        pts += ",".join([str(i) for i in vertices[vert]]) + " "
        color_rgba=mapper.to_rgba(nrank[ri])
        color = mpl.colors.rgb2hex(color_rgba)
    svgpol += '<a href="#c' + str(ri) + '" rel="modal:open"><polygon points="' + pts.strip() + '"  fill="' + color + '" stroke="black" stroke-width="0.2%" /></a>\n'
    vidbox += '<div class="modal" id="c' + str(ri) + '">\n'
    sids = dfpoints['sid'][ri]
    for el in sids.split(','):
        vidbox += '<video width="200" height="150" controls preload="none" poster="http://lablita.it/app/actionsim/png/' + el + '.png"><source src="http://lablita.it/app/actionsim/mp4/' + el + '.mp4" type="video/mp4" /></video>\n'
    vidbox += "</div>"
    ri += 1

for r in range(len(vor.point_region)):
  dfpoints.at[r,'x'] = vor.points[r][0]
  dfpoints.at[r,'y'] = vor.points[r][1]
plt.axis('off')
plt.savefig('tsne2d.png', dpi=600, bbox_inches='tight',pad_inches = 0)
dfpoints.to_csv('points.csv')

# html
text_file = open("map.html", "w")
text_file.write('''<html><head>
                <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.0.0/jquery.min.js"></script>
                <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-modal/0.9.1/jquery.modal.min.js"></script>
                <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/jquery-modal/0.9.1/jquery.modal.min.css" />
                </head><body>\n''')
text_file.write(vidbox)
text_file.write('<svg viewBox="' + str(xmin) + ' ' + str (ymin) + ' ' + str(xmax+abs(xmin)) + ' ' + str(ymax+abs(ymin)) + '" style="position:fixed; top:0; left:0; height:100%; width:100%;" preserveAspectRatio="none" xmlns="http://www.w3.org/2000/svg">\n')
text_file.write(svgpol)
text_file.write("</body></html>")
text_file.close()
